refactor(test): hoist duplicated instant-query wire helper into test/spec/wire
promql_test.go and exotic_test.go each carried their own copy of "issue an
HTTP instant PromQL query and decode the response into a property.Outcome".
Hoist the shared logic into wire.RunInstant, parameterized by
InstantOptions.AllowScalar and .ExtraEscapeChars to preserve each site's
exact prior behavior (exotic's matrix needs scalar decoding and a wider
escape set; property's generator only ever produces vector results, so it
keeps erroring on anything else). Repoint the other test/property call
sites (instant_window_test.go, rate_dup_timestamp_test.go, logql_test.go,
traceql_test.go) that depended on the now-removed package-private helpers
to the new exported wire.RunInstant / wire.EscapeQuery / wire.ParseSample.

Pure refactor, no behavioral change.
